Refactor usable sample filter
