+++ src/app/page.tsx
@@ -67,7 +67,13 @@ const Home = async () => {
         mx='-7.3vw'
         borderBottom='10px solid var(--chakra-colors-diamond-500)'
       >
-        <HStack bg='rgba(0, 5, 77, 0.7)' px='5vw' backdropFilter='blur(5px)' py='15vh' flexWrap="wrap">
+        <HStack
+          bg='rgba(0, 5, 77, 0.7)'
+          px='5vw'
+          backdropFilter='blur(5px)'
+          py='15vh'
+          flexWrap='wrap'
+        >
           <VStack flex='1 0 0' alignItems='start'>
             <Heading>SCAUP</Heading>
             <Heading size='md' fontWeight='200'>
@@ -75,7 +81,7 @@ const Home = async () => {
               arametrisation
             </Heading>
           </VStack>
-          <HStack flexWrap="wrap">
+          <HStack flexWrap='wrap'>
             <InfoBox title='Calendar' href={`${process.env.PATO_URL}/calendar`}>
               View calendar
             </InfoBox>
