+++ src/app/proposals/[proposalId]/sessions/[visitNumber]/shipments/[shipmentId]/(view)/[itemType]/[itemId]/layoutContent.test.tsx
@@ -1,4 +1,4 @@
-import { TreeData } from "@/components/visualisation/treeView";
+import { TreeData } from "@/types/forms";
 import { BaseShipmentItem } from "@/mappings/pages";
 import { puck, renderWithProviders, testInitialState } from "@/utils/test-utils";
 import { fireEvent, screen, waitFor } from "@testing-library/react";
