+++ src/app/proposals/[proposalId]/sessions/[visitNumber]/shipments/[shipmentId]/(view)/layoutContent.test.tsx
@@ -1,4 +1,4 @@
-import { TreeData } from "@/components/visualisation/treeView";
+import { TreeData } from "@/types/forms";
 import { initialState } from "@/features/shipment/shipmentSlice";
 import { getCurrentStepIndex } from "@/mappings/pages";
 import { renderWithProviders, sample, testInitialState } from "@/utils/test-utils";
