+++ src/components/containers/Generic.tsx
@@ -1,6 +1,6 @@
 import { BaseContainerProps, useChildLocationManager } from "@/components/containers";
 import { ChildSelector } from "@/components/containers/ChildSelector";
-import { TreeData } from "@/components/visualisation/treeView";
+import { TreeData } from "@/types/forms";
 import { selectActiveItem } from "@/features/shipment/shipmentSlice";
 import { BaseShipmentItem, Step } from "@/mappings/pages";
 import { Box, Button, Heading, List, ListItem, Text, useDisclosure } from "@chakra-ui/react";
