+++ src/components/containers/Puck.tsx
@@ -1,5 +1,5 @@
 import { ChildSelector } from "@/components/containers/ChildSelector";
-import { TreeData } from "@/components/visualisation/treeView";
+import { TreeData } from "@/types/forms";
 import { selectActiveItem } from "@/features/shipment/shipmentSlice";
 import { PositionedItem } from "@/mappings/forms/sample";
 import { BaseShipmentItem } from "@/mappings/pages";
